Squashed 'public/legacy/' changes from 584c594b6c..3391119ebf
3391119ebf SuiteCRM 7.15.2 Release
26a852b622 Update composer packages
a65e63674f Remove post-filter from getMessagesFromDate() to prevent timezone-related message loss
49864b8995 Add composite uid and mailbox_id index to improve dedup query times
1428721eff Fix ArgumentCountError in ImapHandler::getHeaderInfo() on PHP 8.1+
5ab58e0c81 Fix IMAP timezone dates ignored on import
ab4be6140f Fix #7687 - Update sample pdf loader to ignore invalid files
8f60471bf5 Update sample select to default to none selected
95f41d8993 Fix PDF template detail display shows off screen
b36d11c541 Fix #3855 - Update PDF template samples
eef10d3249 Fix #10874 issue with Associated user not linking correctly
83185bcd06 Fix #10763 - SMTP credentials wrong after changing a field in outbound email settings
475c84095a Fix #10859 - Auth window loses connection to the opener when requesting an OAuth token
0c3fb1ce9b Add optimised indexes for target list joins
c13edb2394 Fix #8989 - Exporting filtering on relate field db failure
785f1519ca Fix #10023 - Fix deleting from imap
90ba9891b2 Fix #10023 - Fix special characters in email folder
b5a7147f9e Fix #10023 - Order child folders alphabetically
3766bc0e11 Fix #10023 - Fix sent folder not creating after first save
7fa60d37c1 Fix #10023 - Delete sent folder if none on save
8145610403 Fix #10023 - add check if not empty mailbox
6765164599 Fix #10023 - Fix emails not saving as seperate folders
dc5fb267dc Fix #10023 - only create sent folder if added
0cefab6dc5 Fix #10023 - Update settings to go to inbound email
d342055a0b Fix #10023 - Add primary folder functionality
81ca1ce4db Fix #10023 - Fix folder selection not showing on drafts
e647bb87cd Fix #10023 - Add primary folder label
57f294c9c3 Fix #10023 - Add primary folder to defs
12aca6facd Fix #10023 - Add primary folder field
962f9c1862 Fix #10023 - Retrieve correct mailbox on folder retrieve
4d4bb4b760 Fix #10023 - Fix delimiter not setting correctly
a2ea8525b4 Fix #10023 - Email Folders Don't Select
ddf8cd668b Fix #10826 - Reporting - Increase length of 'value' field to accommodate large values
b7f42fd9a3 Fix #10826 - Workflows - Increase length of 'value' field to accommodate large values
df64be9364 Fix #10594 Fix issue with relationship errors filling the logs
9ab278d3db Fix #10594 Fix issue with relationship errors filling the logs
0cba3bd247 Fix #10646 - Fix basic module name not saving on different languages
51b0a18a3d Fix #10646 - Fix language files being overwritten when loading new modules
58ecc4bf6e Fix #10646 - Fix setting all app list strings on new modules
c3cef9a4b2 Fix #10830 - Primary email doesn't update when importing
47fe52ceb1 Fix #10849 - Update composer dependencies
989863e5b1 Fix #9788 - Database failures in subpanels
d205dbc483 Fix salesagility#10499 - Error when importing in the Locations module
7a02f43280 Fix #10756 - From set incorrectly
18ca1e8e07 Fix #8951 - Fix rich text field validation
12df22ccc6 Fix #8289 - Fix Import data for floating point fields wrongly saved
41f7137752 Fix #10477 - Do not exclude email sends from the message queue without filtering by campaign

git-subtree-dir: public/legacy
git-subtree-split: 3391119ebfd7811d36054d2a4a7150d251625773

+++ Api/V8/Service/RelationshipService.php
@@ -2,6 +2,7 @@
 namespace Api\V8\Service;
 
 use Api\V8\BeanDecorator\BeanManager;
+use Api\V8\Helper\ModuleAccessChecker;
 use Api\V8\JsonApi\Helper\AttributeObjectHelper;
 use Api\V8\JsonApi\Helper\PaginationObjectHelper;
 use Api\V8\JsonApi\Response\DataResponse;
@@ -17,6 +18,7 @@ use Slim\Http\Request;
 use \SugarBean;
 use \DomainException;
 use SuiteCRM\Exception\AccessDeniedException;
+use SuiteCRM\Exception\NotAllowedException;
 
 #[\AllowDynamicProperties]
 class RelationshipService
@@ -37,15 +39,26 @@ class RelationshipService
     protected $paginationHelper;
 
     /**
+     * @var ModuleAccessChecker
+     */
+    protected $moduleAccessChecker;
+
+    /**
      * @param BeanManager $beanManager
      * @param AttributeObjectHelper $attributeHelper
      * @param PaginationObjectHelper $paginationHelper
+     * @param ModuleAccessChecker $moduleAccessChecker
      */
-    public function __construct(BeanManager $beanManager, AttributeObjectHelper $attributeHelper, PaginationObjectHelper $paginationHelper)
-    {
+    public function __construct(
+        BeanManager $beanManager,
+        AttributeObjectHelper $attributeHelper,
+        PaginationObjectHelper $paginationHelper,
+        ModuleAccessChecker $moduleAccessChecker
+    ) {
         $this->beanManager = $beanManager;
         $this->attributeHelper = $attributeHelper;
         $this->paginationHelper = $paginationHelper;
+        $this->moduleAccessChecker = $moduleAccessChecker;
     }
 
     /**
@@ -57,6 +70,10 @@ class RelationshipService
     {
         $response = new DocumentResponse();
         $sourceBean = $params->getSourceBean();
+        $this->assertSelfOrAdmin($sourceBean->module_dir, $sourceBean->id);
+        if ($sourceBean->module_dir === 'Employees') {
+            $this->assertAdmin($sourceBean->module_dir);
+        }
 
         if (!$sourceBean->ACLAccess('view') || !$sourceBean->ACLAccess('list')) {
             throw new AccessDeniedException();
@@ -90,6 +107,20 @@ class RelationshipService
             $data = [];
             /** @var SugarBean $relatedBean */
             foreach ($relatedBeans as $relatedBean) {
+
+                try {
+                    $this->moduleAccessChecker->checkAccess($relatedBean->module_dir);
+                } catch (NotAllowedException $exception) {
+                    continue;
+                }
+
+                if ($relatedBean->module_dir === 'Employees') {
+                    try {
+                        $this->assertAdmin($relatedBean->module_dir);
+                    } catch (AccessDeniedException $exception) {
+                        continue;
+                    }
+                }
 
                 if (!$relatedBean->ACLAccess('view') || !$relatedBean->ACLAccess('list')) {
                     continue;
@@ -132,12 +163,15 @@ class RelationshipService
     public function createRelationship(CreateRelationshipParams $params)
     {
         $sourceBean = $params->getSourceBean();
+        $this->assertAdmin($sourceBean->module_dir);
 
-        if (!$sourceBean->ACLAccess('view') || !$sourceBean->ACLAccess('list')) {
+        if (!$sourceBean->ACLAccess('view') || !$sourceBean->ACLAccess('edit') || !$sourceBean->ACLAccess('list')) {
             throw new AccessDeniedException();
         }
 
         $relatedBean = $params->getRelatedBean();
+        $this->moduleAccessChecker->checkAccess($relatedBean->module_dir);
+        $this->assertAdmin($relatedBean->module_dir);
 
         if (!$relatedBean->ACLAccess('view') || !$relatedBean->ACLAccess('list')) {
             throw new AccessDeniedException();
@@ -171,12 +205,15 @@ class RelationshipService
     public function createRelationshipByLink(CreateRelationshipByLinkParams $params)
     {
         $sourceBean = $params->getSourceBean();
+        $this->assertAdmin($sourceBean->module_dir);
 
-        if (!$sourceBean->ACLAccess('view') || !$sourceBean->ACLAccess('list')) {
+        if (!$sourceBean->ACLAccess('view') || !$sourceBean->ACLAccess('edit') || !$sourceBean->ACLAccess('list')) {
             throw new AccessDeniedException();
         }
 
         $relatedBean = $params->getRelatedBean();
+        $this->moduleAccessChecker->checkAccess($relatedBean->module_dir);
+        $this->assertAdmin($relatedBean->module_dir);
 
         if (!$relatedBean->ACLAccess('view') || !$relatedBean->ACLAccess('list')) {
             throw new AccessDeniedException();
@@ -226,6 +263,7 @@ class RelationshipService
     public function deleteRelationship(DeleteRelationshipParams $params)
     {
         $sourceBean = $params->getSourceBean();
+        $this->assertAdmin($sourceBean->module_dir);
 
         if (!$sourceBean->ACLAccess('view') || !$sourceBean->ACLAccess('edit') || !$sourceBean->ACLAccess('list')) {
             throw new AccessDeniedException();
@@ -233,14 +271,6 @@ class RelationshipService
 
         $linkFieldName = $params->getLinkedFieldName();
         $relatedBeans = $sourceBean->get_linked_beans($linkFieldName);
-
-        if (!empty($relatedBeans)) {
-            $firstRelatedBean = reset($relatedBeans);
-            if (!$firstRelatedBean->ACLAccess('view') || !$firstRelatedBean->ACLAccess('list')) {
-                throw new AccessDeniedException();
-            }
-        }
-
         $relatedBeanId = $params->getRelatedBeanId();
 
         $relatedBean = array_filter($relatedBeans, function ($bean) use ($relatedBeanId) {
@@ -258,6 +288,13 @@ class RelationshipService
         }
 
         $relatedBean = array_shift($relatedBean);
+        $this->moduleAccessChecker->checkAccess($relatedBean->module_dir);
+        $this->assertAdmin($relatedBean->module_dir);
+
+        if (!$relatedBean->ACLAccess('view') || !$relatedBean->ACLAccess('list')) {
+            throw new AccessDeniedException();
+        }
+
         $this->beanManager->deleteRelationshipSafe($sourceBean, $relatedBean, $linkFieldName);
 
         $response = new DocumentResponse();
@@ -274,5 +311,32 @@ class RelationshipService
         ));
 
         return $response;
+    }
+
+    /**
+     * @param string $module
+     * @throws AccessDeniedException
+     */
+    private function assertAdmin(string $module): void
+    {
+        global $current_user;
+
+        if (in_array($module, ['Users', 'Employees'], true) && !$current_user->isAdmin()) {
+            throw new AccessDeniedException();
+        }
+    }
+
+    /**
+     * @param string $module
+     * @param string $id
+     * @throws AccessDeniedException
+     */
+    private function assertSelfOrAdmin(string $module, string $id): void
+    {
+        global $current_user;
+
+        if ($module === 'Users' && !$current_user->isAdmin() && $id !== $current_user->id) {
+            throw new AccessDeniedException();
+        }
     }
 }
